revert to double

+++ app/src/main/java/com/omouravictor/invest_view/presenter/wallet/model/AssetBySearchUiModel.kt
@@ -10,5 +10,5 @@ data class AssetBySearchUiModel(
     val type: String,
     val region: String,
     val currency: String,
-    var price: Float = 0f
+    var price: Double = 0.0
 ) : Parcelable
